// src/store/enums/StoreEnums.ts
enum Actions {
  // action types
  ADD_BODY_CLASSNAME = "addBodyClassName",
  REMOVE_BODY_CLASSNAME = "removeBodyClassName",
  ADD_BODY_ATTRIBUTE = "addBodyAttribute",
  REMOVE_BODY_ATTRIBUTE = "removeBodyAttribute",
  ADD_CLASSNAME = "addClassName",
  VERIFY_AUTH = "verifyAuth",
  LOGIN = "login",
  LOGOUT = "logout",
  REGISTER = "register",
  UPDATE_USER = "updateUser",
  FORGOT_PASSWORD = "forgotPassword",
  SET_BREADCRUMB_ACTION = "setBreadcrumbAction",
  SET_THEME_MODE_ACTION = "setThemeModeAction",

  // GLOBAL
  FETCH_REGIONS = "fetchRegions",
  FETCH_CONDITION_LIST = "fetchConditionTypes",
  FETCH_CONDITION_TYPE_LIST = "fetchConditionTypeList",

  // Participant
  FETCH_PARTICIPANT_SINGLE = "fetchParticipantSingle",

  // PROGRAM
  FETCH_PROGRAMS = "fetchPrograms",
  FETCH_PARTICIPANT_LIST = "fetchParticipantList",
}

enum Mutations {
  // mutation types
  SET_CLASSNAME_BY_POSITION = "appendBreadcrumb",
  PURGE_AUTH = "logOut",
  SET_AUTH = "setAuth",
  SET_USER = "setUser",
  SET_PASSWORD = "setPassword",
  SET_ERROR = "setError",
  SET_BREADCRUMB_MUTATION = "setBreadcrumbMutation",
  SET_LAYOUT_CONFIG = "setLayoutConfig",
  RESET_LAYOUT_CONFIG = "resetLayoutConfig",
  OVERRIDE_LAYOUT_CONFIG = "overrideLayoutConfig",
  OVERRIDE_PAGE_LAYOUT_CONFIG = "overridePageLayoutConfig",
  SET_THEME_MODE_MUTATION = "setThemeModeMutation",

  // GLOBAL
  SET_REGIONS = "setRegions",
  SET_CONDITION_TYPES = "setConditionTypes",
  SET_CONDITIONS = "setConditions",

  // Participant
  SET_PARTICIPANT_SINGLE = "setParticipantSingle",
  SET_PARTICIPANT_STATISTICS = "setParticipantStatistics",

  // PROGRAM
  SET_PROGRAMS = "setPrograms",
  SET_PARTICIPANT_LIST = "setParticipantList",
}

export { Actions, Mutations };